refactor package structure for Event, EventPersonResponse, EventTest, and Person classes
