+++ src/main/java/com/keyin/eventsystem/event/EventPersonResponse.java
@@ -1,4 +1,4 @@
-package com.keyin.eventsystem;
+package com.keyin.eventsystem.event;
 
 public class EventPersonResponse {
     private String eventName;
