+++ src/test/java/com/keyin/eventsystem/event/EventTest.java
@@ -1,4 +1,4 @@
-package com.keyin.eventsystem;
+package com.keyin.eventsystem.event;
 
 public class EventTest {
